Start implementing tests
The jobrunner is now able to build Python code that executes tests,
based on the JSON string containing tests stored with the question. In
principle this works when previewing the question, although the Python
code doesn't import question libraries yet (nor do question libraries
currently exist...) so therefore the result is a Python exception.

Next steps:

* Write some simple tests in a test library.

* Make sure that this test library is imported (and present in JOBE
  somehow).

* Interpret the output given by the Python code.

And later:

* Build a system (like CodeRunner prototypes) to handle undirected and
  directed graphs separately.