Add support for answer types
There are now three answer types: undirected, directed, and fsm. The user
can select one of these in the question form, and then the correct list
of checks is shown for that type. At quiz time, also the correct module
is loaded for actually running the checks.

Remaining problem: reload the question editing form when the answer type
is modified. In this case the list of questions, answer, and preload
will all need to be erased.

Fixes issue #9.