Ensure that feedback is always shown
In Moodle the quiz behavior determines whether feedback is visible or
not. In GraphChecker questions (just like in CodeRunner questions) it
does not make sense to not show feedback. Therefore in CodeRunner the
corresponding method is overridden to ensure feedback is always shown
(or actually, shown or hidden depending on the question settings).

In GraphChecker this override was removed because we also removed the
corresponding question setting. This commit ensures that the feedback
is now always shown.

Fixes issue #42.