Implement tester tool for UI plugins and checks
This allows developers to test their checks without actually having to
run a Moodle server.

This commit moves the preprocessing functions out of the Twig template
(because in Python, we cannot execute Twig templates). This has not been
changed in the PHP code, so at the moment, running tests in Moodle will
fail.