Created a simple help button inside the toolbar
Added a base class 'Button' to easily create other buttons from later on.

Created a simple help button (with no functionality yet) in the toolbar (WIP).

Also found out that Ctrl+Alt+Shift+M works (for me) to see the JSON-encoded version of the graph and checks in the tester tool.