Replaced the help button, with a button in the toolbar
Took longer than expected.

The HelpButton object 'extends' the Button object, which has an onClick() function. The toolbar has an event listener, and calls the 'onClick()' function of the according button.

When the HelpButton is clicked, an overlay appears on the screen, similar to the 'Add Checks' dialog in the tester tool. It has one issue: when scrolling the graph UI, the overlay stays on the same position on the screen (possibly due to the use of 'position: absolute'. I still have to fix this)