Added distinction between petri net places and transitions
To run the tester tool for petri nets, use the command:
python tester.py petri

In the graph editor, double click to create a new node. This node is a 'place' by default. In order to convert between 'places' and 'transitions' double click the desired node.

Furthermore, the graph saves and reloads the JSON graph encoding correctly for the created petri nets.