Allow failing in the preprocess method
In the preprocessor, if it is determined that the input graph has some
basic error in it because of which it does not make sense to even run
the checks, it is now allowed to raise an exception. This will then be
properly shown in the interface (with the exception message being shown
as well).

As for grading, this counts just like failing a normal check (i.e., the
grade is counted, with no marks).