Overwrite CSS rules from Essential
To make our custom styles work properly (hopefully... I cannot test with
Essential here).

Unfortunately this is a massive hack; we need to use !important to
overwrite the very specific rules set by Essential. It may be possible
to do this in a slightly more proper way, but since I cannot test with
Essential, and I have to do this without looking at the result, doing it
in a crude way is unavoidable for now.