Fix icons still being white with Essential
I was now able to actually test by uploading the stylesheet to my
webserver and forcibly inserting it into the page header in Firefox. It
now seems to look fine.

I think I also found a bug / weird behavior of the Firefox Inspector in
case an !important CSS rule is overwritten by a non-!important rule on a
sub-element. In this case, the rule on the sub-element prevails, but the
!important one is still being shown (with the other rule striked out).

What do you mean, frustrated? I'm not frustrated at all!