Clear form fields when changing the answer type
And also properly restore the old answer type when the user clicks
Cancel.

The next step will be to change the UI parameters after the answer type
has been changed.

This also fixes an issue in the checks UI where it would crash when
given the empty string. (Empty string is not JSON, but it should still
accept it and turn it in an empty array of checks.)