Finishing Petri net requirements
Contains the following:
Fix for the help display not appearing.
Restricting links in Petri nets to be between places and transitions (or vice versa).
Drawing arrows to square of Petri net a bit nicer (does not work perfectly in all circumstances yet).
Simple version for switching to draw mode when pressing CTRL (no CSS styling of buttons yet, and does not remember the selected object).