Milestone Draw mode improvements
This commit incorporates the changes for this milestone. It does not include positioning the Petri labels intelligently yet.

Furthermore, in this commit, when the label is focused on (automatically or manually by the user) and the user presses 'Enter' or 'Escape' the label gets de-focused. In my opinion this improves the workflow when quickly switching from 'Select' mode to 'CTRL-Draw' mode. It can be changed/removed if needed.
'Escape' also deselects all objects.