Changes based on last meeting (24 July 2020)
Changes:
When the label is focussed, pressing CTRL changes the mode to draw mode.
A spinner to set the link label text in Petri nets is now available.
When the label text in a node is greater than the node, the text is drawn outside the node.
When the text is inside the node, and the node has a dark colour (black, blue, green, purple), the text is drawn in white.
Disabled text selection in colour menu dropdown.
For Petri net place nodes, and when the text does not fit inside the node, the text is now placed on the bottom, right, top, or left of the node, where possible.
Improved draw mode shadows for petri net transitions.
Made moving the selection smoother (due to snapping issues).
When vertices and edges have the same colours available, a dropdown menu is shown when of both type are selected.