Fix incorrectly modified null checks
I changed "== null" checks into "=== null" in the previous commit
(because these result in lint warnings). But it turns out that the check
was actually testing for undefined, not null, so "=== null" returns
false then. Therefore, just made it "if (object)" instead of
"if (object == null)" now, to fix the issue and better indicate the
intention.