Extra changes: fixing some bugs, and rewriting the input paremeters
In undirected graphs, there are no invisible self loops anymore.
For petri place nodes, the label is now always outside of the node.
Other (non-default) color codes have been implemented. Also created a Color enum and implemented that.
Rewrote the input parameters. The only ones which are not active at the moment are: vertex_labels, edge_labels, and single_initial_state (for FSMs).