Basic undo/redo functionality
The undo/redo functionality of the graph works as intented. However, when typing labels, it saves the graph for each extra letter typed of a label. This should still be changed to only save the graph when the entire label is filled in.

Furthermore, also fixed small bug of still displaying FSM options when switching to draw mode