Regexes for vertices and edges
Added regexes for vertices and edges. Furthermore, fixed a small bug with the undo-redo, where the toolbar selection options were not updated/displayed properly

Note: You can add regex values, without start and end signs (i.e. forward-slashes /), for the fields "vertex_label_regex" and "edge_label_regex" in the types.json file to enable regexes for vertices/edges.
An example of this is: "vertex_label_regex": "^[a-zA-Z]+$". In this regex only letters are allowed