Implement graph editor for the checks UI
Unfortunately for some reason the graph UI, while being perfectly
functional, does not draw anything on the canvas. The canvas just stays
while, but it is actually possible to draw something, you just need to
do it without seeing what you are doing.

Furthermore the parameters are still hardcoded and need to be derived
from the parameters set in the other graph UIs on the page.

See issue #117.