Fixed "Menu scaling"
Fixed the graph toolbar 'wrapping' to create more space for the button options. Although still not entirely sure if this is intuitive (perhaps a fixed toolbar height, instead of a dynamically changing one, will work better when using the UI).
Also the initial graph height is now set to 350px for more space initially.
Removed the 40px gutter too.
When entering the graphUI there is checked if the control button is pressed. This is done to prevent a bug where the control button was released after the user clicked outside the graphUI, causing the status of the control button to not get updated.