Fix checkboxes not working in dialogs
After an hours-long debugging session, partially together with Arthur,
we figured out that if you do `return false;` in jQuery, it calls
preventDefault() on the event. And apparently, this cancels default
events also on the children, causing the checkboxes not to change state
anymore when clicked.

In our case this happened because in our dialog code, we put an event
handler with `return false;` on the dialog to avoid click events
bubbling upwards to the backdrop (which has an on-click-handler to close
the dialog that obviously should not be triggered if someone clicks
inside the dialog). This commit replaces this `return false;` by a call to
stopPropagation(), which is enough to stop the bubbling to the backdrop,
but does not cancel the default event.