Implement locked nodes and edges
A node / edge with 'locked': true in its JSON specification will be
locked, that is, it will be drawn with a gray shadow, and it will not be
possible for the user to edit or delete it, other than moving the
object.

This is implemented by making it impossible to select a locked object.
That unfortunately also makes it hard to use multiselect to move several
locked objects at the same time.