Add edit_locked mode used for the preload editor
The new edit_locked input parameter of the Graph UI does two things.

* It ignores locked parameters in the input graph, so locked vertices
  and edges are editable and not shown in gray.

* In the output JSON, everything is always marked locked.

This is now used for the preload editor, so it is possible to have a
preloaded graph with vertices and edges that are locked for the student,
but the teacher can still edit it in the preload editor.