Show error if checking code crashes outside checks
It used to be the case that we'd show exception stack traces if the
Python checker functions would raise an exception. However, we ran into
an issue where an import statement raised an exception, and in that case
no stack trace was shown. This commit rectifies that behavior.

Fixes issue #109.