Refactor the jobrunner code
This is in preparation of supporting partial grades. This refactor makes
sure that we hand the JSON string from the checks UI straight to the
Python code, without PHP having to parse it first (except we still need
to parse it to find out which modules we need to pass to the sandbox,
but that doesn't need to care about partial grades).

See issue #128.