Finish implementing partial grades
This takes the grade computed by the checker code and properly handles
it on the Moodle side.

The UI now asks teachers to input a percentage of grades for each
partial grade block. That is because it seems hard to implement absolute
grades there (what to do if the number of points for the question
changes?)

Closes issue #128.