Signal output on stdout in the tester
If checks print to stdout, this prevents them from working properly on
the server. This is because the server reads the output produced by
the checker code to figure out if checks passed or failed. This involves
parsing the output to JSON, which of course fails if the checker code
itself prints random things to stdout.

The tester, until now, allowed printing to stdout, however. So we had a
few instances where checks worked in the tester, were added to the
repository, and then turned out to fail on the server because debug
print statements were left in.

This commit teaches the tester that if a check produces stdout output,
it should just reject it (done internally by raising an exception).

Fixes issue #119.