Validate that all requested checks actually exist
When trying to save a question whose checks specification contains
checks that don't exist, we now show an error message instead of saving
the question anyway and failing at runtime.

In practice this shouldn't happen because the checks UI won't allow
users to even add non-existing checks, but firstly, Ctrl+Alt+M could be
used to edit the raw JSON to add them, and secondly, someone may remove
checks from the server after the question was already made.

This also generates a validation error if the JSON is invalid in the
first place.

See issue #13.