Expose Moodle's penalty option in the edit form
Background: CodeRunner had a rather complicated penalty regime field,
where a teacher could enter things like "20,40,60,70,80,...", to define
how much a student should be penalized if they get the right answer
after making 1, 2, ... mistakes first. We found this too complicated for
our purposes, so we removed support for this. Hence, students would
always get full points, even if they needed many tries.

However, in practice it does turn out to be useful to be able to deduct
points per try. Instead of reinstating CodeRunner's penalty regime, we
opted instead to use Moodle's built-in support for penalties in adaptive
mode. Therefore, this commit just exposes this field in the question
edit form, and that is all that is needed.

Note that this solution depends on the fact that the
adaptive_adapted_for_coderunner behavior contains a fallback: if the
question doesn't define a penalty regime, it just calls the parent class
(which is the standard adaptive behavior, which implements Moodle's
penalty scheme). If this fallback ever gets removed, we should fork the
behavior.

See also the CodeRunner commit [1] in which Moodle's penalty scheme was
replaced by the penalty regimes.

Fixes issue #123.

[1] https://github.com/trampgeek/moodle-qtype_coderunner/commit/087aa778a4a7f23a50cad6d9baede891587701ba