Fetch UI params for answertype on the client side
Before this commit, on loading the question edit form, the PHP code
would determine the current answertype and based on that, set the
correct UI params (for sample answer and preload) and list of checks
(for the checks UI).

While this worked fine in general, it caused a problem whenever a
validation error would occur. After such an error, Moodle reloads the
form, but the form data is not saved. Therefore, if a user changes the
answertype from A to B and then saves, and if the validation then fails,
then

* the form is loaded with B in the answertype field;
* answertype A is still stored in the database.

In other words, in this situation we'd initialize the UI params and list
of checks with answertype A, while answertype B would be shown in the
UI. That's a recipe for disaster, especially because the user may
already have picked some checks (for answertype B) before submitting the
form which failed validation. In that case, the checks UI will now try
to show these checks as belonging to answertype A, which of course
fails.

I spent quite a long time researching if it is possible in Moodle's form
framework to obtain the actual answertype value B that is shown in the
field, but this seems impossible as at the time our form-construction
code runs, Moodle still ‘thinks’ that A is going into the answertype
field. Only later, Moodle apparently ‘realizes’ that it should overwrite
the value by B. (I figured this out by putting debug statements in
Moodle's form code, in particular in the setValue() method of the select
element.)

As a different solution, this commit simply does not try to determine
the answertype on the PHP side when constructing the form. Instead, in
authorform.js we simply check the type on pageload, and fetch the
necessary checks list and UI params asynchronously. Note that we were
already doing this if the user changed the answertype field. Therefore,
we can simply reuse this code.

Fixes issue #149.