Small refactoring of ui_graph.js
Placed global variables in separate file.
Placed two 'classes' (GraphCanvas and HelpOverlay) in separate files. I could not manage to place them both inside a subfolder, as those files could not be found in the tester tool in that case.
Placed code documentation (comments) above some of the functions. To be continued.