More changes
Fixed a bug where nothing was allowed, and highlighting is pressed. It gave an error that the 'undo' and 'redo' button could not be found.

Changed 'add_vertex' and 'delete_vertex' to 'edit_vertex', and likewise for 'add_edge' and 'delete_edge', and changed according handling code. Also changed this in the question.php code (hopefully it is done correctly!)

Fixed that pressing the delete button removed locked items. Now it does not anymore (and thus works correctly).

Also fixed a small 'grunt amd' error