Fixed FSM initial state handling
The StartLink is now not an object in the JSON string anymore, but is instead created, for a node, by setting 'initial = true' on the node.

Start links are not selectable and editable anymore. When highlighting a node, it's start link (if it has one) is also highlighted. Start links currently only take on the default edge color value, not the color value of their according node (yet).

Also rewrote/simplified the script for the placements of start links, as it previously apparently did not work well when dealing with dynamically moving edges. It now divides the circle in 16 parts, and chooses the one nearest to the top-left, with the most surrounding space.