Fix graph editor receiving the incorrect type
This is a problem for answertypes where the type passed as a UI param to
the graph editor differs from the name of the answertype itself. It was
never the intention to send the answertype name instead of the type
specified in the ui_params for the answertype. However, until now for
all of our question types, these two were the same, so we never noticed
the problem.