Overhaul graph editor to use add and move modes
Instead of select and draw, as we had before. The real change is that we
allow selecting objects in the draw mode, call that mode "add mode"
instead, and make it the default mode. This means that students that are
going to draw a graph (which is probably the most common action) can do
so immediately, without changing modes.

This is an experimental change for now; we need to test with all graph
types and parameter settings to see if everything works properly.