Graph editor: avoid regenerating toolbar buttons
Currently, every time the user selects an object in the graph editor,
because of which the toolbar needs to be updated, this is implemented by
clearing out the entire toolbar and recreating the relevant controls.
This is unnecessary and introduces bugs.

For example, issue #231 (editor forgets to generate an undo state for a
label change, if the user clicks on another object after editing the
label) is caused by the fact that the label edit field, which is
supposed to have a focus out event inserting the undo state, is removed
from the toolbar altogether before this focus out event can trigger.

This commit reworks the toolbar code to just hide fields that are
irrelevant for the current selection, and show them again when needed.
This way, fields do not need to be recreated all the time.

Note that this code is still buggy and needs to be developed further.